Refactor Activity model to remove inherited notification callbacks (#1396)
* Refactor Activity model to remove inherited notification callbacks. Migrate column type

* Don't create activity for test requests

* Make deprecated models read only

* Remove erroneous index

* Allow specs to create meetings and test requests to ensure support for legacy data
