+++ app/controllers/comments_controller.rb
@@ -2,7 +2,7 @@ class CommentsController < ApplicationController
   before_action :set_investigation
 
   def create
-    @comment = Activity.new(comment_activity_params)
+    @comment = CommentActivity.new(comment_activity_params)
     @comment.investigation = @investigation
     @comment.source = UserSource.new(user: current_user)
     @investigation = @investigation.decorate
@@ -23,7 +23,7 @@ class CommentsController < ApplicationController
   end
 
   def new
-    @comment = @investigation.activities.new
+    @comment = CommentActivity.new
     @investigation = @investigation.decorate
   end
