+++ app/models/activity.rb
@@ -3,8 +3,6 @@ class Activity < ApplicationRecord
 
   has_one :source, as: :sourceable, dependent: :destroy
 
-  after_save :notify_relevant_users
-
   def attached_image?
     nil
   end
@@ -26,7 +24,7 @@ class Activity < ApplicationRecord
     "#{subtitle_slug} by #{source&.show(viewer)}, #{pretty_date_stamp}"
   end
 
-  def search_index;  end
+  def search_index; end
 
   def self.sanitize_text(text)
     return text.to_s.strip.gsub(/[*_~]/) { |match| "\\#{match}" } if text
@@ -40,31 +38,10 @@ class Activity < ApplicationRecord
     # where necessary should be implemented by subclasses
   end
 
-  def email_update_text(viewer = nil); end
-
-  def email_subject_text
-    "#{investigation.case_type.upcase_first} updated"
-  end
-
   # Used to determine which view template to use for new records with metadata
   # instead of pre-generated HTML
   def template_name
     self.class.name.delete_prefix("AuditActivity::").underscore
-  end
-
-  def entities_to_notify
-    entities = users_to_notify
-
-    teams_to_notify.each do |team|
-      if team.team_recipient_email.present?
-        entities << team
-      else
-        users_from_team = team.users.active
-        entities.concat(users_from_team)
-      end
-    end
-
-    entities.uniq
   end
 
 private
@@ -74,25 +51,4 @@ private
   end
 
   def subtitle_slug; end
-
-  def notify_relevant_users
-    entities_to_notify.each do |entity|
-      email = entity.is_a?(Team) ? entity.team_recipient_email : entity.email
-      NotifyMailer.investigation_updated(investigation.pretty_id, entity.name, email, email_update_text(entity), email_subject_text).deliver_later
-    end
-  end
-
-  def users_to_notify
-    return [] unless investigation.owner.is_a? User
-    return [] if source&.user == investigation.owner
-
-    [investigation.owner]
-  end
-
-  def teams_to_notify
-    return [] unless investigation.owner.is_a? Team
-    return [] if source&.user&.team == investigation.owner
-
-    [investigation.owner]
-  end
 end
