+++ app/models/audit_activity/correspondence/add_meeting.rb
@@ -4,8 +4,8 @@ class AuditActivity::Correspondence::AddMeeting < AuditActivity::Correspondence:
   include ActivityAttachable
   with_attachments transcript: "transcript", related_attachment: "related attachment"
 
-  def self.from(*)
-    raise "Deprecated - no longer supported"
+  def readonly?
+    true
   end
 
   def activity_type
@@ -14,10 +14,6 @@ class AuditActivity::Correspondence::AddMeeting < AuditActivity::Correspondence:
 
   def restricted_title(_user)
     "Meeting added"
-  end
-
-  def email_update_text(viewer = nil)
-    "Meeting details added to the #{investigation.case_type.upcase_first} by #{source&.show(viewer)}."
   end
 
 private
