+++ app/models/audit_activity/investigation/change_notifying_country.rb
@@ -1,8 +1,4 @@
 class AuditActivity::Investigation::ChangeNotifyingCountry < AuditActivity::Investigation::Base
-  def self.from(*)
-    raise "Deprecated - use ChangeNotifyingCountry.call instead"
-  end
-
   def self.build_metadata(investigation)
     updated_values = investigation.previous_changes.slice(:notifying_country)
 
@@ -26,9 +22,4 @@ class AuditActivity::Investigation::ChangeNotifyingCountry < AuditActivity::Inve
   def new_country
     metadata["updates"]["notifying_country"].second
   end
-
-private
-
-  # Do not send investigation_updated mail. This is handled by the ChangeNotifyingCountry service
-  def notify_relevant_users; end
 end
