+++ app/models/audit_activity/investigation/change_safety_and_compliance_data.rb
@@ -1,8 +1,4 @@
 class AuditActivity::Investigation::ChangeSafetyAndComplianceData < AuditActivity::Investigation::Base
-  def self.from(*)
-    raise "Deprecated - use ChangeSafetyAndComplianceData.call instead"
-  end
-
   def self.build_metadata(investigation)
     updated_values = investigation.previous_changes.slice(:hazard_type, :hazard_description, :non_compliant_reason, :reported_reason)
 
@@ -14,9 +10,4 @@ class AuditActivity::Investigation::ChangeSafetyAndComplianceData < AuditActivit
   def title(*)
     "Safety and compliance status changed"
   end
-
-private
-
-  # Do not send investigation_updated mail. This is handled by the ChangeSafetyAndComplianceData service
-  def notify_relevant_users; end
 end
