+++ app/models/audit_activity/investigation/team_added.rb
@@ -14,11 +14,4 @@ class AuditActivity::Investigation::TeamAdded < AuditActivity::Investigation::Ba
   def team
     Team.find(metadata["team"]["id"])
   end
-
-private
-
-  # This is handled by the AddTeamToCase service, but this override is required
-  # to prevent a duplicate investigation_updated email being enqueued, which
-  # will fail
-  def notify_relevant_users; end
 end
