+++ app/models/audit_activity/investigation/team_deleted.rb
@@ -12,11 +12,4 @@ class AuditActivity::Investigation::TeamDeleted < AuditActivity::Investigation::
   def team
     Team.find(metadata["team"]["id"])
   end
-
-private
-
-  # This is handled by RemoveTeamFromCase, but this override is required to
-  # prevent a duplicate investigation_updated email being enqueued, which will
-  # fail
-  def notify_relevant_users; end
 end
