+++ app/models/audit_activity/investigation/update_visibility.rb
@@ -1,8 +1,4 @@
 class AuditActivity::Investigation::UpdateVisibility < AuditActivity::Investigation::Base
-  def self.from(_investigation)
-    raise "Deprecated - use ChangeCaseVisibility.call instead"
-  end
-
   def self.build_metadata(investigation, rationale)
     updated_values = investigation.previous_changes.slice(:is_private)
 
@@ -25,9 +21,6 @@ private
 
     { "updates" => { "is_private" => [nil, title.match?(/\s+restricted$/im)] } }
   end
-
-  # Do not send investigation_updated mail. This is handled by the ChangeCaseVisibility service
-  def notify_relevant_users; end
 
   def already_in_new_format?
     self[:metadata]&.key?("updates")
