+++ app/models/audit_activity/product/add.rb
@@ -1,8 +1,4 @@
 class AuditActivity::Product::Add < AuditActivity::Product::Base
-  def self.from(*)
-    raise "Deprecated - use AddProductToCase.call instead"
-  end
-
 private
 
   def subtitle_slug
