+++ app/models/audit_activity/test/request.rb
@@ -1,15 +1,7 @@
+# Recording of test requests is deprecated - existing data is still supported
 class AuditActivity::Test::Request < AuditActivity::Test::Base
-  def self.from(test)
-    title = "Test requested: #{test.product.name}"
-    super(test, title)
-  end
-
-  def self.date_label
-    "Date requested"
-  end
-
-  def email_update_text(viewer = nil)
-    "Test request was added to the #{investigation.case_type} by #{source&.show(viewer)}."
+  def readonly?
+    true
   end
 
 private
