+++ app/models/test/request.rb
@@ -1,10 +1,5 @@
+# Recording of test requests is deprecated - existing data is still supported
 class Test::Request < Test
-  after_create :create_audit_activity
-
-  def create_audit_activity
-    AuditActivity::Test::Request.from(self)
-  end
-
   def pretty_name
     "testing request"
   end
