+++ db/migrate/20210520120140_change_activity_type_column.rb
@@ -1,0 +1,9 @@
+class ChangeActivityTypeColumn < ActiveRecord::Migration[6.1]
+  def change
+    safety_assured do
+      change_column_default :activities, :type, from: "CommentActivity", to: nil
+      change_column_null :activities, :type, false
+      add_index :activities, :type
+    end
+  end
+end
